+++ aws/container-linux/kubernetes/workers.tf
@@ -10,6 +10,7 @@ module "workers" {
   instance_type   = "${var.worker_type}"
   os_channel      = "${var.os_channel}"
   disk_size       = "${var.disk_size}"
+  spot_price      = "${var.worker_price}"
 
   # configuration
   kubeconfig            = "${module.bootkube.kubeconfig}"